fixing merging errors (please be careful when moving stuff around so as not to break the tools)
